Edits to the on page documentation for the tutorial 14
I'm not sure if the description of the behaviours is quite correct
where it is referring to 'session keys are copied first to the new
window'. When a new window is created, from the link the session key
is 'null', but the storage key is loaded from the previous value. This
behaviour is unsurprising, but diffent to the description, so I have
taken the liberty of changing the description so that it matches the
observed behaviour.